+++ HyMD/hybrid/src/main/java/de/hpi/is/md/hybrid/impl/preprocessed/CompressedRelation.java
@@ -26,7 +26,7 @@ class CompressedRelation {
 	}
 
 	<T> int indexOf(Column<T> column) {
-		return columnIndexes.get(column);
+		return columnIndexes.get(column).intValue();
 	}
 
 	int size() {
