+++ HyMD/sim/src/test/java/de/hpi/is/md/sim/CachedSimilarityMeasureTest.java
@@ -19,14 +19,14 @@ public class CachedSimilarityMeasureTest {
 
 	@Test
 	public void test() {
-		doReturn(0.5).when(sim).calculateSimilarity(UnorderedPair.of(1, 2));
-		doReturn(1.0 / 3).when(sim).calculateSimilarity(UnorderedPair.of(1, 3));
+		doReturn(Double.valueOf(0.5)).when(sim).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(2)));
+		doReturn(Double.valueOf(1.0 / 3)).when(sim).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(3)));
 		SimilarityMeasure<Integer> cachedSim = CachedSimilarityMeasure.<Integer>builder()
 			.similarityMeasure(sim)
 			.build();
-		assertThat(cachedSim.calculateSimilarity(1, 2)).isEqualTo(0.5);
-		assertThat(cachedSim.calculateSimilarity(2, 1)).isEqualTo(0.5);
-		assertThat(cachedSim.calculateSimilarity(1, 3)).isEqualTo(1.0 / 3);
+		assertThat(cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(2))).isEqualTo(0.5);
+		assertThat(cachedSim.calculateSimilarity(Integer.valueOf(2), Integer.valueOf(1))).isEqualTo(0.5);
+		assertThat(cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(3))).isEqualTo(1.0 / 3);
 	}
 
 	@Test
@@ -35,10 +35,10 @@ public class CachedSimilarityMeasureTest {
 			.similarityMeasure(sim)
 			.maximumSize(1)
 			.build();
-		cachedSim.calculateSimilarity(1, 2);
-		verify(sim).calculateSimilarity(UnorderedPair.of(1, 2));
-		cachedSim.calculateSimilarity(1, 2);
-		verify(sim).calculateSimilarity(UnorderedPair.of(1, 2));
+		cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(2));
+		verify(sim).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(2)));
+		cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(2));
+		verify(sim).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(2)));
 	}
 
 	@Test
@@ -47,10 +47,10 @@ public class CachedSimilarityMeasureTest {
 			.similarityMeasure(sim)
 			.maximumSize(0)
 			.build();
-		cachedSim.calculateSimilarity(1, 2);
-		verify(sim).calculateSimilarity(UnorderedPair.of(1, 2));
-		cachedSim.calculateSimilarity(1, 2);
-		verify(sim, times(2)).calculateSimilarity(UnorderedPair.of(1, 2));
+		cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(2));
+		verify(sim).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(2)));
+		cachedSim.calculateSimilarity(Integer.valueOf(1), Integer.valueOf(2));
+		verify(sim, times(2)).calculateSimilarity(UnorderedPair.of(Integer.valueOf(1), Integer.valueOf(2)));
 	}
 
 }
