+++ dfd/dfdAlgorithm/src/fdiscovery/pruning/PruneTable.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import fdiscovery.columns.ColumnCollection;
 
 // from rhs to lhs
-public abstract class PruneTable extends HashMap<ColumnCollection, HashMap<Long, ArrayList<ColumnCollection>>> {
+public abstract class PruneTable extends HashMap<ColumnCollection, HashMap<Integer, ArrayList<ColumnCollection>>> {
 
 	private static final long serialVersionUID = 4470955427882698208L;
 	
@@ -23,14 +23,14 @@ public abstract class PruneTable extends HashMap<ColumnCollection, HashMap<Long,
 	
 	public void addValue(ColumnCollection RHS, ColumnCollection LHS) {
 		if (!this.containsKey(RHS)) {
-			this.put(RHS, new HashMap<Long, ArrayList<ColumnCollection>>());
+			this.put(RHS, new HashMap<Integer, ArrayList<ColumnCollection>>());
 		}
-		if (!this.get(RHS).containsKey(LHS.cardinality())) {
-			this.get(RHS).put(LHS.cardinality(), new ArrayList<ColumnCollection>());
+		if (!this.get(RHS).containsKey(Integer.valueOf(LHS.cardinality()))) {
+			this.get(RHS).put(Integer.valueOf(LHS.cardinality()), new ArrayList<ColumnCollection>());
 		}
 //		System.out.println(this.get(RHS));
 //		System.out.println(String.format("Column:\t%s\t%d", LHS, LHS.cardinality()));
-		ArrayList<ColumnCollection> dependencies = this.get(RHS).get(LHS.cardinality());
+		ArrayList<ColumnCollection> dependencies = this.get(RHS).get(Integer.valueOf(LHS.cardinality()));
 //		System.out.println(dependencies);
 		dependencies.add(LHS);
 	}
